+++ MultipleTextEditor/Form1.cs
@@ -65,17 +65,7 @@ namespace MultipleTextEditor
         }
         
 
-        private void FontToolStripMenuItem_Click(Object sender, EventArgs e)
-        {
-            ChangeToolStripMenuItemBackgroundColors();
-            // フォントダイアログを表示してユーザーにフォントサイズを選択させる
-            FontDialog fontDialog = new FontDialog();
-            if (fontDialog.ShowDialog() == DialogResult.OK)
-            {
-                // 選択されたフォントサイズを適用する
-                text_memo.Font = fontDialog.Font;
-            }
-        }
+        
 
         private void timer1_Tick_1(object sender, EventArgs e)
         {
